refactor: delete old callable function directories and extract shared utilities
Remove old read-profile/, write-profile/, and create-profile/ callable function
implementations now that they've been fully replaced by Elysia REST API.

Extract markdown serialization utilities to profiles-api for reuse across
GitHub service operations.

Changes:
- Delete functions/src/read-profile/ (replaced by GET /api/profiles/me)
- Delete functions/src/write-profile/ (replaced by PUT /api/profiles/me)
- Delete functions/src/create-profile/ (replaced by POST /api/profiles/me)
- Create profiles-api/utils/markdown-serialization.ts
- Extract serializeToMarkdown() and parseExistingMetadata()
- Update imports in GitHub services to use new location
- Fix unused import in check-slug-available.test.ts

All tests pass: 51 backend tests, 18 e2e tests
